+++ src/gfx_text.cpp
@@ -119,7 +119,7 @@ NO_DISCARD bool R_CreateTextureAtlas(INOUT FontDef& f) {
 // ============================================================================
 
 void R_DrawText(
-    int localClientNum, OPTIONAL_IN const FontDef* font, std::string_view text,
+    size_t localClientNum, OPTIONAL_IN const FontDef* font, std::string_view text,
     float x, float y, float xscale, float yscale, const glm::vec3& color,
     bool right
 ) {
@@ -257,7 +257,7 @@ void R_DrawText(
 
 std::array<std::array<GfxTextDraw, 256>, MAX_LOCAL_CLIENTS> r_textDraws;
 
-bool R_GetTextDraw(int localClientNum, size_t id, OUT GfxTextDraw*& draw) {
+bool R_GetTextDraw(size_t localClientNum, size_t id, OUT GfxTextDraw*& draw) {
     draw = nullptr;
     assert(id < r_textDraws.at(localClientNum).size());
 
@@ -267,7 +267,7 @@ bool R_GetTextDraw(int localClientNum, size_t id, OUT GfxTextDraw*& draw) {
 }
 
 
-bool R_FindFreeTextDraw(int localClientNum, OUT size_t& index, OUT GfxTextDraw*& draw) {
+bool R_FindFreeTextDraw(size_t localClientNum, OUT size_t& index, OUT GfxTextDraw*& draw) {
     index = (size_t)-1;
     draw = nullptr;
 
@@ -283,7 +283,7 @@ bool R_FindFreeTextDraw(int localClientNum, OUT size_t& index, OUT GfxTextDraw*&
 }
 
 bool R_AddTextDraw(
-    int localClientNum, FontDef* font, std::string text, float x, float y,
+    size_t localClientNum, FontDef* font, std::string text, float x, float y,
     float xscale, float yscale, glm::vec3 color, bool active, bool right,
     OUT size_t& id
 ) {
@@ -305,7 +305,7 @@ bool R_AddTextDraw(
     return true;
 }
 
-bool R_UpdateTextDraw(int localClientNum, size_t id, std::string text) {
+bool R_UpdateTextDraw(size_t localClientNum, size_t id, std::string text) {
     GfxTextDraw* d = nullptr;
     if (!R_GetTextDraw(localClientNum, id, d))
         return false;
@@ -314,7 +314,7 @@ bool R_UpdateTextDraw(int localClientNum, size_t id, std::string text) {
     return true;
 }
 
-bool R_ActivateTextDraw(int localClientNum, size_t id, bool active) {
+bool R_ActivateTextDraw(size_t localClientNum, size_t id, bool active) {
     GfxTextDraw* d = nullptr;
     R_GetTextDraw(localClientNum, id, d);
     bool wasActive = d->active;
@@ -322,7 +322,7 @@ bool R_ActivateTextDraw(int localClientNum, size_t id, bool active) {
     return wasActive;
 }
 
-bool R_RemoveTextDraw(int localClientNum, size_t id) {
+bool R_RemoveTextDraw(size_t localClientNum, size_t id) {
     GfxTextDraw* d = nullptr;
     if (!R_GetTextDraw(localClientNum, id, d))
         return false;
@@ -335,14 +335,14 @@ bool R_RemoveTextDraw(int localClientNum, size_t id) {
     return true;
 }
 
-void R_ClearTextDraws(int localClientNum) {
+void R_ClearTextDraws(size_t localClientNum) {
     for (auto& c : r_textDraws.at(localClientNum)) {
         c.active = false;
         c.free = true;
     }
 }
 
-void R_DrawTextDraws(int localClientNum) {
+void R_DrawTextDraws(size_t localClientNum) {
     for (const auto& c : r_textDraws.at(localClientNum)) {
         if (!c.free && c.active) {
             R_DrawText(
